requirements.txt
streamlit>=1.42
pandas>=2.2
plotly>=5.24
openpyxl>=3.1
python-dotenv>=1.0
pydantic>=2.7
sqlalchemy>=2.0
openai>=1.68
langchain>=0.3
langchain-openai>=0.3
